Removed shell=True
